% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_grid.R
\docType{data}
\name{gdp_grid}
\alias{gdp_grid}
\title{Gridded GDPs}
\format{An object of class \code{RasterStack} of dimension 287 x 720 x 31.}
\source{
\url{https://github.com/Nowosad/global_population_and_gdp}
}
\usage{
gdp_grid
}
\description{
A dataset containing gridded GDPs (unit: PPP, Billion US$2005/yr) for years 1980-2100.
Values from 1980-2010 were estimated by downscaling country-level data,
while those in 2020-2100 are estimated by downscaling country-level data
from SSP1, SSP2, and SSP3 scenarios.
}
\keyword{datasets}
