% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_grid.R
\docType{data}
\name{pop_grid}
\alias{pop_grid}
\title{Gridded Populations}
\format{An object of class \code{RasterStack} of dimension 287 x 720 x 31.}
\source{
\url{https://github.com/Nowosad/global_population_and_gdp}
}
\usage{
pop_grid
}
\description{
A dataset containing gridded populations (unit: millions) for years 1980-2100.
Values from 1980-2010 were estimated by downscaling country-level data,
while those in 2020-2100 are estimated by downscaling country-level data
from SSP1, SSP2, and SSP3 scenarios.
}
\keyword{datasets}
